git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.MW/Dev/Features/ActivityMigration/BLFlex;C24977

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/Shared/Formatter/IWordPluralizer.cs
@@ -2,6 +2,6 @@
 {
     public interface IWordPluralizer
     {
-        string GetPluralFor(int numberToRead);
+        string GetPluralFor(long numberToRead);
     }
 }
